organize signup structure
